Generate Random Board State and Pretty Print it
Generate a random board state where each cell of the state is either
alive or dead (1 or 0). State is a two layer matrix of cells. Each cell
is either 0 or 1. If a randomly generated number is below a certain
threshold, the cell is initialized to dead. Above the threshold, and the
cell is initialized to alive.

Pretty print the current state through the render function, where a
living cell is represented by '#' and a dead cell is a space.