Create unit tests for get-neighbours function
Create and use unit tests for the get-neighbours function used in the
next-state function. Tests include a small state (3x3) and the framework
for larger state tests.

At the moment, the tests focus on ensuring the get-neighbours function
sucessfully gathers the neighbourhood of the selected cell, including
itself. Special focus is given to the board-wrapping ability of
get-neighbours.